He empezado a integrar la API de OpenAI para transcribir los sets a partir de texto / voz. Clase Transcriber creada y en proceso de terminar la clase de TextParser.

+++ src/training_sessions/domain/models.py
@@ -41,7 +41,7 @@ class Set:
     mean_velocity: Optional[float]= None
     peak_velocity: Optional[float]= None
     power: Optional[float]= None
-    rir: Optional[float]= None
+    rir: Optional[int]= None
